split visualization into script per type of plot
